+++ src/Model/Validation/PageValidator.php
@@ -12,38 +12,11 @@
  */
 namespace MeCms\Model\Validation;
 
-use MeCms\Validation\AppValidator;
+use MeCms\Validation\PageAndPostValidator;
 
 /**
  * Page validator class
  */
-class PageValidator extends AppValidator
+class PageValidator extends PageAndPostValidator
 {
-    /**
-     * Construct.
-     *
-     * Adds some validation rules.
-     * @uses MeCms\Validation\AppValidator::__construct()
-     */
-    public function __construct()
-    {
-        parent::__construct();
-
-        //Category
-        $this->add('category_id', [
-            'naturalNumber' => [
-                'message' => I18N_SELECT_VALID_OPTION,
-                'rule' => 'naturalNumber',
-            ],
-        ])->requirePresence('category_id', 'create');
-
-        //Title
-        $this->requirePresence('title', 'create');
-
-        //Slug
-        $this->requirePresence('slug', 'create');
-
-        //Text
-        $this->requirePresence('text', 'create');
-    }
 }