+++ src/Validation/PageAndPostValidator.php
@@ -9,15 +9,18 @@
  * @copyright   Copyright (c) Mirko Pagliai
  * @link        https://github.com/mirko-pagliai/me-cms
  * @license     https://opensource.org/licenses/mit-license.php MIT License
+ * @since       2.26.6
  */
-namespace MeCms\Model\Validation;
+namespace MeCms\Validation;
 
-use MeCms\Validation\BannerAndPhotoValidator;
+use MeCms\Validation\AppValidator;
 
 /**
- * Photo validator class
+ * Abstract class for `PageValidator` and `PostValidator` classes.
+ *
+ * This class provides some methods and properties common to both classes.
  */
-class PhotoValidator extends BannerAndPhotoValidator
+abstract class PageAndPostValidator extends AppValidator
 {
     /**
      * Construct.
@@ -29,12 +32,21 @@ class PhotoValidator extends BannerAndPhotoValidator
     {
         parent::__construct();
 
-        //Album
-        $this->add('album_id', [
+        //Category
+        $this->add('category_id', [
             'naturalNumber' => [
                 'message' => I18N_SELECT_VALID_OPTION,
                 'rule' => 'naturalNumber',
             ],
-        ])->requirePresence('album_id', 'create');
+        ])->requirePresence('category_id', 'create');
+
+        //Title
+        $this->requirePresence('title', 'create');
+
+        //Slug
+        $this->requirePresence('slug', 'create');
+
+        //Text
+        $this->requirePresence('text', 'create');
     }
 }